requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
requests==2.31.0
user-agents==2.2.0
geoip2==4.8.0
python-dotenv==1.0.0
SQLAlchemy==2.0.25 